refactor command line flags
We were re-using flags from chronoctl-core which made it somewhat hard
to extend. It also couples transport and client/APIs with flags which we
mostly don't use.

This is a partial fork of the chronoctl flags struct with everything we
don't use removed. Having control of this will make it a bit easier to
extend.