refactor outgoing transport for auth context
The way we handle auth is a little awkward. We reuse the swagger
transport from chronoctl for both swagger requests and non-swagger
requests (prom). We also forward the auth token by extracting the bearer
token from a dummy request and then forwarding that to

This attempts to clean that up by partially forking the transport from
chronoctl to allow us to separate out a base round tripper used for both
prom and swagger and creating an auth forwarding round tripper
middleware that allows us to greatly reduce the number of places we have
to deal with the auth token.