refactor outgoing transport for auth context (#32)
The way we handle auth is a little awkward. We reuse the swagger
transport from chronoctl for both swagger requests and non-swagger
requests (prom). We also forward the auth token by extracting the bearer
token from a dummy request and then set an api-token header which 
might be a legacy chronoctl thing. 

This attempts to clean that up by partially forking the transport from
chronoctl to allow us to separate out a base round tripper used for both
prom and swagger and creating an auth forwarding round tripper
middleware that will allows us to greatly reduce the number of places we
have
to deal with the auth token.