rename to chronosphere-mcp (#40)
If you use many mcp-servers, cloning this repo as "mcp-server" makes it
hard to distinguish what it is. Renaming to chronosphere-mcp should
help. This will be a breaking change on everyone's laptops, so better do
this early before too many people has it checked out.

This is a breaking change for anyone who has cloned this repo. They'll
need to change .git/config and/or go.mod to point at the new address
https://github.com/chronosphereio/chronosphere-mcp