add tool schema validation test (#45)
It's too easy to define an invalid schema for Array and Object types.
Cursor for example is strict about each tool input having a well defined
schema so a missing "items" type on the array will cause all too
requests to fail.

This adds a schema validation test for tool schemas.