simply server start up
Move configuration to yaml files with optional commandline arguments
that can override the yaml file values so command line flags are
completely optional when configs are set in confg.yaml.

Removed the `--api-token-filename` option as we already have 3 other
ways to provide the token (config, env var, cmd flag).

Removed the `--api-url` option since it can now be set in the config
file.