feat: add ability to limit log query results
This tool often returns too much data and blow up the context window
when directly passed back to an LLM. This allows it to specify a limit
and offset to return a subset of the log data to avoid context
exhaustion. The API doesn't support pagination, so having an offset
gives some ability to read the "next page" of logs, however, it doesnt'
work when timestamps used are relative (e.g. start -1h)