fix: encode trace & span ID response as hex
By default the generated spec encodes the []byte array traceID and
spanID as a base64 string which isn't recognized by our tracing system.
This changes the encoding to hex.

We have to re-define a lot of generated fields to make this work, and
using the new json v2's "unknown field" feature to retain any fields we
didn't define ourselves in the output.